Refactor "wait loops"
Put the common code for the check-sleep loops in a separate module, and
make timeouts and sleep intervals optional, using a common default value.

Also introduce a new `TimeoutError` to make error handling easier.
