+++ fishtank/src/cluster.ts
@@ -8,7 +8,8 @@ import { join, resolve } from 'path'
 import { Docker, Labels } from './backend'
 import { getConfig } from './config'
 import * as naming from './naming'
-import { DEFAULT_WAIT_TIMEOUT, INTERNAL_RPC_TCP_PORT, Node } from './node'
+import { INTERNAL_RPC_TCP_PORT, Node } from './node'
+import { DEFAULT_WAIT_TIMEOUT } from './waitLoop'
 
 export const DEFAULT_BOOTSTRAP_NODE_NAME = 'bootstrap'
 export const CLUSTER_LABEL = 'fishtank.cluster'
